+++ aiootp/ciphers/cipher_interface.py
@@ -130,9 +130,18 @@ class CipherInterface(FrozenInstance):
                 Thus it may simplify implementing efficient features, such
                 as search or routing, though care must still be taken when
                 considering how leaking such metadata may be harmful.
-                Keeping this value constant is strongly discouraged, though
-                the salt misuse-reuse resistance of the cipher extends up
-                to ~256**(len(iv)/2 + len(siv_key)/2) encryptions/second.
+
+                Keeping this value constant is strongly discouraged. Though,
+                the cipher's salt misuse-reuse resistance is ruled by the
+                combination of the automatically incorporated `timestamp`,
+                `iv`, & `siv_key`. The risk calculation starts with setting
+                r = len(iv + siv_key) / 3. Then, all else staying constant,
+                once 256**r messages are encrypted within a second, each
+                additional encrypted message within that same second begins
+                to have more than a 256**(-r) chance of generating a repeat
+                context.
+
+                See: https://github.com/rmlibre/aiootp/issues/16
 
         `aad`: An arbitrary bytes value that a user decides to categorize
                 keystreams. It's authenticated as associated data & safely
@@ -173,9 +182,18 @@ class CipherInterface(FrozenInstance):
                 Thus it may simplify implementing efficient features, such
                 as search or routing, though care must still be taken when
                 considering how leaking such metadata may be harmful.
-                Keeping this value constant is strongly discouraged, though
-                the salt misuse-reuse resistance of the cipher extends up
-                to ~256**(len(iv)/2 + len(siv_key)/2) encryptions/second.
+
+                Keeping this value constant is strongly discouraged. Though,
+                the cipher's salt misuse-reuse resistance is ruled by the
+                combination of the automatically incorporated `timestamp`,
+                `iv`, & `siv_key`. The risk calculation starts with setting
+                r = len(iv + siv_key) / 3. Then, all else staying constant,
+                once 256**r messages are encrypted within a second, each
+                additional encrypted message within that same second begins
+                to have more than a 256**(-r) chance of generating a repeat
+                context.
+
+                See: https://github.com/rmlibre/aiootp/issues/16
 
         `aad`: An arbitrary bytes value that a user decides to categorize
                 keystreams. It's authenticated as associated data & safely
@@ -283,9 +301,18 @@ class CipherInterface(FrozenInstance):
                 Thus it may simplify implementing efficient features, such
                 as search or routing, though care must still be taken when
                 considering how leaking such metadata may be harmful.
-                Keeping this value constant is strongly discouraged, though
-                the salt misuse-reuse resistance of the cipher extends up
-                to ~256**(len(iv)/2 + len(siv_key)/2) encryptions/second.
+
+                Keeping this value constant is strongly discouraged. Though,
+                the cipher's salt misuse-reuse resistance is ruled by the
+                combination of the automatically incorporated `timestamp`,
+                `iv`, & `siv_key`. The risk calculation starts with setting
+                r = len(iv + siv_key) / 3. Then, all else staying constant,
+                once 256**r messages are encrypted within a second, each
+                additional encrypted message within that same second begins
+                to have more than a 256**(-r) chance of generating a repeat
+                context.
+
+                See: https://github.com/rmlibre/aiootp/issues/16
 
         `aad`: An arbitrary bytes value that a user decides to categorize
                 keystreams. It's authenticated as associated data & safely
@@ -313,9 +340,18 @@ class CipherInterface(FrozenInstance):
                 Thus it may simplify implementing efficient features, such
                 as search or routing, though care must still be taken when
                 considering how leaking such metadata may be harmful.
-                Keeping this value constant is strongly discouraged, though
-                the salt misuse-reuse resistance of the cipher extends up
-                to ~256**(len(iv)/2 + len(siv_key)/2) encryptions/second.
+
+                Keeping this value constant is strongly discouraged. Though,
+                the cipher's salt misuse-reuse resistance is ruled by the
+                combination of the automatically incorporated `timestamp`,
+                `iv`, & `siv_key`. The risk calculation starts with setting
+                r = len(iv + siv_key) / 3. Then, all else staying constant,
+                once 256**r messages are encrypted within a second, each
+                additional encrypted message within that same second begins
+                to have more than a 256**(-r) chance of generating a repeat
+                context.
+
+                See: https://github.com/rmlibre/aiootp/issues/16
 
         `aad`: An arbitrary bytes value that a user decides to categorize
                 keystreams. It's authenticated as associated data & safely
@@ -471,9 +507,18 @@ class CipherInterface(FrozenInstance):
                 Thus it may simplify implementing efficient features, such
                 as search or routing, though care must still be taken when
                 considering how leaking such metadata may be harmful.
-                Keeping this value constant is strongly discouraged, though
-                the salt misuse-reuse resistance of the cipher extends up
-                to ~256**(len(iv)/2 + len(siv_key)/2) encryptions/second.
+
+                Keeping this value constant is strongly discouraged. Though,
+                the cipher's salt misuse-reuse resistance is ruled by the
+                combination of the automatically incorporated `timestamp`,
+                `iv`, & `siv_key`. The risk calculation starts with setting
+                r = len(iv + siv_key) / 3. Then, all else staying constant,
+                once 256**r messages are encrypted within a second, each
+                additional encrypted message within that same second begins
+                to have more than a 256**(-r) chance of generating a repeat
+                context.
+
+                See: https://github.com/rmlibre/aiootp/issues/16
 
         `aad`: An arbitrary bytes value that a user decides to categorize
                 keystreams. It's authenticated as associated data & safely
@@ -493,9 +538,18 @@ class CipherInterface(FrozenInstance):
                 Thus it may simplify implementing efficient features, such
                 as search or routing, though care must still be taken when
                 considering how leaking such metadata may be harmful.
-                Keeping this value constant is strongly discouraged, though
-                the salt misuse-reuse resistance of the cipher extends up
-                to ~256**(len(iv)/2 + len(siv_key)/2) encryptions/second.
+
+                Keeping this value constant is strongly discouraged. Though,
+                the cipher's salt misuse-reuse resistance is ruled by the
+                combination of the automatically incorporated `timestamp`,
+                `iv`, & `siv_key`. The risk calculation starts with setting
+                r = len(iv + siv_key) / 3. Then, all else staying constant,
+                once 256**r messages are encrypted within a second, each
+                additional encrypted message within that same second begins
+                to have more than a 256**(-r) chance of generating a repeat
+                context.
+
+                See: https://github.com/rmlibre/aiootp/issues/16
 
         `aad`: An arbitrary bytes value that a user decides to categorize
                 keystreams. It's authenticated as associated data & safely
@@ -521,9 +575,18 @@ class CipherInterface(FrozenInstance):
                 Thus it may simplify implementing efficient features, such
                 as search or routing, though care must still be taken when
                 considering how leaking such metadata may be harmful.
-                Keeping this value constant is strongly discouraged, though
-                the salt misuse-reuse resistance of the cipher extends up
-                to ~256**(len(iv)/2 + len(siv_key)/2) encryptions/second.
+
+                Keeping this value constant is strongly discouraged. Though,
+                the cipher's salt misuse-reuse resistance is ruled by the
+                combination of the automatically incorporated `timestamp`,
+                `iv`, & `siv_key`. The risk calculation starts with setting
+                r = len(iv + siv_key) / 3. Then, all else staying constant,
+                once 256**r messages are encrypted within a second, each
+                additional encrypted message within that same second begins
+                to have more than a 256**(-r) chance of generating a repeat
+                context.
+
+                See: https://github.com/rmlibre/aiootp/issues/16
 
         `aad`: An arbitrary bytes value that a user decides to categorize
                 keystreams. It's authenticated as associated data & safely
@@ -531,9 +594,7 @@ class CipherInterface(FrozenInstance):
                 each permutation of `key`, `salt`, & `iv`.
 
         `iv`: An ephemeral, uniform, random value that's generated by
-                the encryption algorithm. Ensures salt misue / reuse
-                security even if the `key`, `salt`, & `aad` are the same
-                for ~256**(len(iv)/2 + len(siv_key)/2) encryptions/second.
+                the encryption algorithm.
 
         `ttl`: An amount of seconds that dictate the allowable age of
                 the decrypted message.
@@ -559,9 +620,18 @@ class CipherInterface(FrozenInstance):
                 Thus it may simplify implementing efficient features, such
                 as search or routing, though care must still be taken when
                 considering how leaking such metadata may be harmful.
-                Keeping this value constant is strongly discouraged, though
-                the salt misuse-reuse resistance of the cipher extends up
-                to ~256**(len(iv)/2 + len(siv_key)/2) encryptions/second.
+
+                Keeping this value constant is strongly discouraged. Though,
+                the cipher's salt misuse-reuse resistance is ruled by the
+                combination of the automatically incorporated `timestamp`,
+                `iv`, & `siv_key`. The risk calculation starts with setting
+                r = len(iv + siv_key) / 3. Then, all else staying constant,
+                once 256**r messages are encrypted within a second, each
+                additional encrypted message within that same second begins
+                to have more than a 256**(-r) chance of generating a repeat
+                context.
+
+                See: https://github.com/rmlibre/aiootp/issues/16
 
         `aad`: An arbitrary bytes value that a user decides to categorize
                 keystreams. It's authenticated as associated data & safely
@@ -569,9 +639,7 @@ class CipherInterface(FrozenInstance):
                 each permutation of `key`, `salt`, & `iv`.
 
         `iv`: An ephemeral, uniform, random value that's generated by
-                the encryption algorithm. Ensures salt misue / reuse
-                security even if the `key`, `salt`, & `aad` are the same
-                for ~256**(len(iv)/2 + len(siv_key)/2) encryptions/second.
+                the encryption algorithm.
 
         `ttl`: An amount of seconds that dictate the allowable age of
                 the decrypted message.